requirements.txt
Django==2.1.4
django-environ==0.4.5

djangorestframework==3.9.0
django-oauth-toolkit==1.2.0
django-cors-headers==2.4.0

django-crispy-forms==1.7.2
django-model-utils==3.1.2
